Distinguish MLB positions with subclasses
In baseball, there are distinct differences in the statistical
categories between field players and pitchers, and these two positions
should be abstracted from one another. For example, if a fielder has
never pitched before, he will never be marked for a balk during a pitch,
so that category is relatively meaningless to the fielder. Therefore,
any categories that have no relevance to a particular position should be
removed from the class.

Signed-Off-By: Robert Clark <robdclark@outlook.com>